chore(size): thin OCR/Merge route shells + manifest-based guard for routes only; enforce main<40KB & route<1KB
