+++ src/ui/copy.ts
@@ -1,0 +1,9 @@
+export const COPY = {
+  result: 'Done',
+  download: 'Download',
+  again: 'Again',
+  chooseFiles: 'Choose files',
+  run: 'Run',
+  ocrTitle: 'OCR',
+  mergeTitle: 'Merge',
+};
