+++ src/ui/mini.tsx
@@ -1,0 +1,5 @@
+import React from 'react';
+export const Row = ({ children, className = '' }: {children: React.ReactNode; className?: string}) =>
+  <div className={`flex items-center gap-2 ${className}`}>{children}</div>;
+export const Btn = ({ onClick, children }: {onClick?: ()=>void; children: React.ReactNode}) =>
+  <button onClick={onClick} className="px-3 py-1 rounded border">{children}</button>;
